requirements.txt
astroid==2.3.3
attrs==19.3.0
cheroot==8.3.0
geojson==2.5.0
isort==4.3.21
jaraco.functools==3.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pathlib==1.0.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
web.py==0.50
wrapt==1.11.2
